+++ pcs/common/tools.py
@@ -4,9 +4,11 @@ from __future__ import (
     print_function,
 )
 
+import sys
 from lxml import etree
 import threading
 
+_PYTHON2 = sys.version[0] == "2"
 
 def simple_cache(func):
     cache = {
@@ -68,4 +70,6 @@ def xml_fromstring(xml):
     # ValueError: Unicode strings with encoding declaration are not supported.
     # Please use bytes input or XML fragments without declaration.
     # So we encode the string to bytes.
-    return etree.fromstring(xml.encode("utf-8"))
+    # In python2 we cannot do that as it causes a UnicodeDecodeError if the xml
+    # contains a non-ascii character.
+    return etree.fromstring(xml if _PYTHON2 else xml.encode("utf-8"))