+++ pcs/lib/commands/test/test_resource_agent.py
@@ -1,3 +1,4 @@
+# coding=utf-8
 from __future__ import (
     absolute_import,
     division,
@@ -8,6 +9,7 @@ import logging
 from lxml import etree
 
 from pcs.test.tools.assertions import assert_raise_library_error, start_tag_error_text
+from pcs.test.tools.command_env import get_env_tools
 from pcs.test.tools.custom_mock import MockLibraryReportProcessor
 from pcs.test.tools.pcs_unittest import mock, TestCase
 
@@ -380,3 +382,85 @@ class TestDescribeAgent(TestCase):
 
         self.assertEqual(len(mock_metadata.mock_calls), 1)
         mock_guess.assert_not_called()
+
+
+class DescribeAgentUtf8(TestCase):
+    def setUp(self):
+        self.env_assist, self.config = get_env_tools(test_case=self)
+        self.config.runner.pcmk.load_agent(
+            agent_filename="resource_agent_ocf_heartbeat_dummy_utf8.xml"
+        )
+
+    def test_describe(self):
+        name = "ocf:heartbeat:Dummy"
+        self.assertEqual(
+            lib.describe_agent(self.env_assist.get_env(), name),
+            {
+                "name": name,
+                "shortdesc": u"Example stateless resource agent: ®",
+                "longdesc": u"This is a Dummy Resource Agent fot testing utf-8"
+                    u" in metadata: ®"
+                ,
+                "parameters": [
+                    {
+                        "advanced": False,
+                        "default": u"/var/run/resource-agents/Dummy-®.state",
+                        "deprecated": False,
+                        "longdesc":
+                            u"Location to store the resource state in: ®",
+                        "name": u"state-®",
+                        "obsoletes": None,
+                        "pcs_deprecated_warning": "",
+                        "required": False,
+                        "shortdesc": u"State file: ®",
+                        "type": "string",
+                    },
+                    {
+                        "advanced": True,
+                        "default": 0,
+                        "deprecated": False,
+                        "longdesc": "Set to 1 to turn on resource agent tracing"
+                            " (expect large output) The trace output will be "
+                            "saved to trace_file, if set, or by default to "
+                            "$HA_VARRUN/ra_trace/<type>/<id>.<action>."
+                            "<timestamp> e.g. $HA_VARRUN/ra_trace/oracle/db."
+                            "start.2012-11-27.08:37:08",
+                        "name": "trace_ra",
+                        "obsoletes": None,
+                        "pcs_deprecated_warning": "",
+                        "required": False,
+                        "shortdesc": "Set to 1 to turn on resource agent "
+                            "tracing (expect large output)",
+                        "type": "integer",
+                    },
+                    {
+                        "advanced": True,
+                        "default": "",
+                        "deprecated": False,
+                        "longdesc": "Path to a file to store resource agent "
+                            "tracing log",
+                        "name": "trace_file",
+                        "obsoletes": None,
+                        "pcs_deprecated_warning": "",
+                        "required": False,
+                        "shortdesc": "Path to a file to store resource agent "
+                            "tracing log",
+                        "type": "string",
+                    }
+                ],
+                "actions": [
+                    {"name": "start", "timeout": "20"},
+                    {"name": "stop", "timeout": "20"},
+                    {"name": "monitor", "interval": "10", "timeout": "20"},
+                    {"name": "meta-data", "timeout": "5"},
+                    {"name": "validate-all", "timeout": "20"},
+                    {"name": u"custom-®", "timeout": "20"},
+                ],
+                "default_actions": [
+                    {"name": "start", "interval": "0s", "timeout": "20"},
+                    {"name": "stop", "interval": "0s", "timeout": "20"},
+                    {"name": "monitor", "interval": "10", "timeout": "20"},
+                    {"name": u"custom-®", "interval": "0s", "timeout": "20"},
+                ],
+            }
+        )