+++ CHANGELOG.md
@@ -12,12 +12,15 @@
 - `resurce update` no longer exits with an error when the `remote-node` meta
   attribute is set to the same value that it already has
   ([rhbz#1502715], [ghissue#145])
+- Listing and describing resource and stonith agents no longer crashes when
+  agents' metadata contain non-ascii characters ([rhbz#1503110])
 
 [ghissue#145]: https://github.com/ClusterLabs/pcs/issues/145
 [rhbz#1230919]: https://bugzilla.redhat.com/show_bug.cgi?id=1230919
 [rhbz#1415197]: https://bugzilla.redhat.com/show_bug.cgi?id=1415197
 [rhbz#1501274]: https://bugzilla.redhat.com/show_bug.cgi?id=1501274
 [rhbz#1502715]: https://bugzilla.redhat.com/show_bug.cgi?id=1502715
+[rhbz#1503110]: https://bugzilla.redhat.com/show_bug.cgi?id=1503110
 
 
 ## [0.9.160] - 2017-10-09